PRODEGY_4_Image-to-Image_Translation_with_cGAN_ProdigyInfoTech/Application/requirements.txt
Flask==2.0.1
tensorflow==2.16.1
Pillow==9.5.0
numpy==1.23.5
matplotlib==3.7.1
Werkzeug==2.2.3 